+++ src/main/java/sunlib/turtle/cache/data/DataCache.java
@@ -11,9 +11,9 @@ public abstract class DataCache implements Cache {
 
     public abstract String get(String key);
 
-    protected abstract void put(String key, String data, int timestamp);
+    protected abstract void put(String key, String data, long timestamp);
 
-    public void put(String key, Object data, int timestamp) {
+    public void put(String key, Object data, long timestamp) {
         put(key, (String) data, timestamp);
     }
 }
